перевод оффсетов камеры на флоаты

+++ main/nvs_templates.h
@@ -5,12 +5,15 @@
 
 #include "esp_system.h"
 
+#include "stdio.h"
+#include "string.h""
+
 bool get_variable(const char* NVS_namespace, const char* NVS_key, int16_t& value);
 
 void get_MPU_offsets_blob(int16_t *offsets);
 void save_MPU_offsets_blob(int16_t offsets[]);
 void restore_MPU_offsets_blob();
 
-void get_OpenMV_offset(int16_t* offset_x, int16_t* offset_y);
-void set_OpenMV_offset(int16_t offset_x, int16_t offset_y);
+void get_OpenMV_offset(float* offset_x, float* offset_y);
+void set_OpenMV_offset(float offset_x, float offset_y);
 void restore_OpenMV_offset();
